backend/requirements.txt
asgiref==3.3.2
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.3
django-extra-fields==3.0.2
django-filter==2.3.0
django-push-notifications==3.0.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==24.3.0
flake8==7.0.0
gunicorn==20.1.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
pillow==10.2.0
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==2.2.1
